+++ delayed_image/delayed_leafs.py
@@ -127,7 +127,7 @@ class DelayedLoad(DelayedImageLeaf):
         >>> stack2 = kwimage.stack_images([stack1, data6], axis=1)
         >>> kwplot.imshow(stack2)
     """
-    def __init__(self, fpath, channels=None, dsize=None, nodata_method=None):
+    def __init__(self, fpath, channels=None, dsize=None, nodata_method=None, num_overviews=None):
         """
         Args:
             fpath (str | PathLike):
@@ -142,10 +142,14 @@ class DelayedLoad(DelayedImageLeaf):
             nodata_method (str | None):
                 How to handle nodata values in the file itself.
                 Can be "auto", "float", or "ma".
+
+            num_overviews (int | None):
+                number of overviews if known a-priori
         """
         super().__init__(channels=channels, dsize=dsize)
         self.meta['fpath'] = fpath
         self.meta['nodata_method'] = nodata_method
+        self.meta['num_overviews'] = num_overviews
         self.lazy_ref = None
 
     @property
@@ -210,12 +214,11 @@ class DelayedLoad(DelayedImageLeaf):
 
     @profile
     def _load_reference(self):
-        nodata_method = self.meta.get('nodata_method', None)
         if self.lazy_ref is None:
             from delayed_image import lazy_loaders
             using_gdal = lazy_loaders.LazyGDalFrameFile.available()
             if using_gdal:
-                # the nodata arg here isn't named that great
+                nodata_method = self.meta.get('nodata_method', None)
                 self.lazy_ref = lazy_loaders.LazyGDalFrameFile(
                     self.fpath, nodata_method=nodata_method)
             else:
@@ -239,6 +242,23 @@ class DelayedLoad(DelayedImageLeaf):
 
     @profile
     def _load_metadata(self):
+        """
+        Ignore:
+            We want to be able to skip the reference load if the metadata is
+            already setup.
+
+            import kwimage
+            from delayed_image.delayed_leafs import *  # NOQA
+            dsize = (32, 32)
+            channels = 'red|green|blue'
+            fpath = kwimage.grab_test_image_fpath('astro', dsize=dsize)
+            self = DelayedLoad(fpath, channels=channels, dsize=dsize, num_overviews=1)
+            xdev.profile_now(self._load_metadata)()
+
+        """
+        required_meta_keys = ('dsize', 'num_channels', 'num_overviews')
+        if all(self.meta[k] is not None for k in required_meta_keys):
+            return self
         self._load_reference()
         if self.lazy_ref is NotImplemented:
             shape = kwimage.load_image_shape(self.fpath)
@@ -285,6 +305,31 @@ class DelayedLoad(DelayedImageLeaf):
             # underlying lazy ref.
             self.lazy_ref.nodata_method = self.meta.get('nodata_method', None)
             return self.lazy_ref
+
+    # Reimplementations of existing properties with specialized logic for speed
+    # @property
+    # def num_channels(self):
+    #     """
+    #     Returns:
+    #         None | int
+    #     """
+    #     return self.meta.get('num_channels', None)
+
+    # @property
+    # def dsize(self):
+    #     """
+    #     Returns:
+    #         None | Tuple[int | None, int | None]
+    #     """
+    #     return self.meta.get('dsize', None)
+
+    # @property
+    # def channels(self):
+    #     """
+    #     Returns:
+    #         None | FusedChannelSpec
+    #     """
+    #     return self.meta.get('channels', None)
 
 
 class DelayedNans(DelayedImageLeaf):