+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.3.1 - Unreleased
 
+### Changed:
+* Minor speed improvements
+* Can now specify `num_overviews` to DelayedLoad to bypass metadata preparation.
+
 
 ## Version 0.3.0 - Released 2024-06-20